+++ ThirdPartyAdapters/liftoffmonetize/liftoffmonetize/src/main/java/com/google/ads/mediation/vungle/VungleConstants.java
@@ -20,5 +20,6 @@ public class VungleConstants {
   public static final String KEY_USER_ID = "userId";
   public static final String KEY_ORIENTATION = "adOrientation";
   public static final String KEY_PLACEMENT_ID = "placementID";
-
+  public static final String KEY_BACK_BUTTON_IMMEDIATELY_ENABLED =
+      "back_button_immediately_enabled";
 }